# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Estacionamento API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/employees": {
            "get": {
                "tags": [
                    "Employee"
                ],
                "summary": "Get all employees",
                "operationId": "7fc06f144c1cd7925c5c391a6609f02d",
                "responses": {
                    "200": {
                        "description": "A list of employees",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GetAllEmployeesResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/employee": {
            "post": {
                "tags": [
                    "Employee"
                ],
                "summary": "Create employee",
                "operationId": "6a4f1e0943433f32bb808e136da78430",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Name of the employee",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "Email address of the employee",
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "description": "Password for the employee account",
                                        "type": "string"
                                    },
                                    "type": {
                                        "description": "Type of employee",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Employees",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CreateEmployeeResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/employee/{id}": {
            "get": {
                "tags": [
                    "Employee"
                ],
                "summary": "Get employee by ID",
                "operationId": "7002cd80d065e00fc43b1794ed7b4d4f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the employee to retrieve",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Employee details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GetEmployeeByIdResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Employee not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Employee"
                ],
                "summary": "Update employee",
                "operationId": "b1f58be54bff0320d2f19085bf535278",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the employee to update",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Name of the employee",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "Email address of the employee",
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "description": "Password for the employee account",
                                        "type": "string"
                                    },
                                    "type": {
                                        "description": "Type of employee",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated employee details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UpdateEmployeeResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Employee not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Employee"
                ],
                "summary": "Delete employee",
                "operationId": "3d61663792ac65f6dbcbf97c9e90e7e2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the employee to delete",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted employee details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DeleteEmployeeByIdResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Employee not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/employee/login": {
            "post": {
                "tags": [
                    "Employee"
                ],
                "summary": "Login employee",
                "operationId": "27a6883919c5b7fbeaafeef58819fd24",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "Email address of the employee",
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "description": "Password for the employee account",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Logged in employee details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LoginEmployeeResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/api/employee/logout": {
            "post": {
                "tags": [
                    "Employee"
                ],
                "summary": "Logout employee",
                "operationId": "fd08e58782056cf92be3740d422354bd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "token": {
                                        "description": "JWT token of the employee",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Logged out successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "CreateEmployeeResource": {
                "title": "Create Employees Resource",
                "description": "Resource for returning a create employee",
                "properties": {
                    "status": {
                        "description": "Indicates if the request was successful or not",
                        "type": "boolean"
                    },
                    "errors": {
                        "description": "List of errors (if any)",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "message": {
                        "description": "Message indicating the status of the response",
                        "type": "string"
                    },
                    "employee": {
                        "description": "Employee",
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "description": "Employee ID",
                                    "type": "integer"
                                },
                                "name": {
                                    "description": "Employee name",
                                    "type": "string"
                                },
                                "email": {
                                    "description": "Employee email",
                                    "type": "string"
                                },
                                "tipo": {
                                    "description": "Employee type",
                                    "type": "integer"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "DeleteEmployeeByIdResource": {
                "title": "Delete Employee Response",
                "description": "Response schema for deleting an employee",
                "properties": {
                    "status": {
                        "description": "Indicates whether the operation was successful",
                        "type": "boolean"
                    },
                    "errors": {
                        "description": "List of errors, if any",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "message": {
                        "description": "Success message if the operation was successful",
                        "type": "string"
                    },
                    "employee": {
                        "description": "Employee information, which is null in this case",
                        "type": "null"
                    }
                },
                "type": "object"
            },
            "GetAllEmployeesResource": {
                "title": "Get All Employees Resource",
                "description": "Resource for returning a list of employees",
                "properties": {
                    "status": {
                        "description": "Indicates if the request was successful or not",
                        "type": "boolean"
                    },
                    "errors": {
                        "description": "List of errors (if any)",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "message": {
                        "description": "Message indicating the status of the response",
                        "type": "string"
                    },
                    "employees": {
                        "description": "List of employees",
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "description": "Employee ID",
                                    "type": "integer"
                                },
                                "name": {
                                    "description": "Employee name",
                                    "type": "string"
                                },
                                "email": {
                                    "description": "Employee email",
                                    "type": "string"
                                },
                                "tipo": {
                                    "description": "Employee type",
                                    "type": "integer"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "GetEmployeeByIdResource": {
                "title": "Get employee Resource",
                "description": "Recourse to return an employee",
                "properties": {
                    "status": {
                        "description": "Indicates if the request was successful or not",
                        "type": "boolean"
                    },
                    "errors": {
                        "description": "List of errors (if any)",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "message": {
                        "description": "Message indicating the status of the response",
                        "type": "string"
                    },
                    "employee": {
                        "description": "Eemployees",
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "description": "Employee ID",
                                    "type": "integer"
                                },
                                "name": {
                                    "description": "Employee name",
                                    "type": "string"
                                },
                                "email": {
                                    "description": "Employee email",
                                    "type": "string"
                                },
                                "tipo": {
                                    "description": "Employee type",
                                    "type": "integer"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "LoginEmployeeResource": {
                "title": "LoginEmployeeResource",
                "description": "Resource para retornar os dados do login de um funcionário",
                "properties": {
                    "status": {
                        "description": "Status do login",
                        "type": "boolean"
                    },
                    "errors": {
                        "description": "Lista de erros",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "message": {
                        "description": "Mensagem de sucesso",
                        "type": "string",
                        "nullable": true
                    },
                    "employee": {
                        "properties": {
                            "id": {
                                "description": "ID do funcionário",
                                "type": "integer"
                            },
                            "token": {
                                "description": "Token de autenticação do funcionário",
                                "type": "string"
                            },
                            "name": {
                                "description": "Nome do funcionário",
                                "type": "string"
                            },
                            "email": {
                                "description": "Email do funcionário",
                                "type": "string"
                            },
                            "tipo": {
                                "description": "Tipo do funcionário",
                                "type": "string"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "UpdateEmployeeResource": {
                "title": "Update Employee Response",
                "description": "Response schema for updating an employee",
                "properties": {
                    "status": {
                        "description": "Indicates whether the operation was successful",
                        "type": "boolean"
                    },
                    "errors": {
                        "description": "List of errors, if any",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "message": {
                        "description": "Success message if the operation was successful",
                        "type": "string"
                    },
                    "employee": {
                        "description": "Details of the updated employee",
                        "properties": {
                            "id": {
                                "description": "ID of the employee",
                                "type": "string"
                            },
                            "name": {
                                "description": "Name of the employee",
                                "type": "string"
                            },
                            "email": {
                                "description": "Email address of the employee",
                                "type": "string",
                                "format": "email"
                            },
                            "tipo": {
                                "description": "Type of employee",
                                "type": "string"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter JWT token to authenticate"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}